🏷️ Ta imot reisedistanse som desimaltall

+++ src/main/kotlin/no/nav/tilleggsstonader/sak/behandling/fakta/BehandlingFaktaDto.kt
@@ -166,7 +166,7 @@ data class FaktaReise(
     val periode: Reiseperiode,
     val dagerPerUke: String,
     val harMerEnn6KmReisevei: JaNei,
-    val lengdeReisevei: Int?,
+    val lengdeReisevei: Double?,
     val harBehovForTransportUavhengigAvReisensLengde: JaNei?,
     val kanReiseMedOffentligTransport: JaNei,
     val offentligTransport: OffentligTransport?,
